Moved transform package to query folder as sqltransformer package
